src/test/java/opennlp/tools/formats/pan/pan16/PAN16ReaderTest.java
/*
 * Licensed to the Apache Software Foundation (ASF) under one or more
 * contributor license agreements.  See the NOTICE file distributed with
 * this work for additional information regarding copyright ownership.
 * The ASF licenses this file to You under the Apache License, Version 2.0
 * (the "License"); you may not use this file except in compliance with
 * the License. You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package opennlp.tools.formats.pan.pan16;

import org.junit.Test;

import java.util.ArrayList;

import static org.junit.Assert.assertEquals;

/**
 * Created by anthony on 6/7/16.
 */
public class PAN16ReaderTest {

  /**
   * Test for PAN16 reader
   * */
  @Test
  public void testReader(){
    ArrayList<PAN16Author> authors = PAN16Reader.getPAN16Authors(0, 19);
    assertEquals("Check the number of authors ", 20, authors.size());
  }


}
